remove badge component
