+++ client/my-sites/plan-features-2023-grid/plan-comparison-grid.tsx
@@ -19,6 +19,7 @@ import { useTranslate } from 'i18n-calypso';
 import { useMemo, useState, useCallback, useEffect, ChangeEvent } from 'react';
 import { useSelector } from 'react-redux';
 import JetpackLogo from 'calypso/components/jetpack-logo';
+import PlanPill from 'calypso/components/plans/plan-pill';
 import { getPlanFeaturesObject } from 'calypso/lib/plans/features-list';
 import PlanTypeSelector, {
 	PlanTypeSelectorProps,
@@ -26,7 +27,6 @@ import PlanTypeSelector, {
 import { getCurrentUserCurrencyCode } from 'calypso/state/currency-code/selectors';
 import PlanFeatures2023GridActions from './actions';
 import PlanFeatures2023GridBillingTimeframe from './billing-timeframe';
-import PopularBadge from './components/popular-badge';
 import PlanFeatures2023GridHeaderPrice from './header-price';
 import { plansBreakSmall, plansBreakLarge } from './media-queries';
 import { Plans2023Tooltip } from './plans-2023-tooltip';
@@ -312,10 +312,7 @@ const PlanComparisonGridHeader: React.FC< PlanComparisonGridHeaderProps > = ( {
 					'plan-comparison-grid__header',
 					getPlanClass( planName ),
 					{
-						'popular-plan-parent-class':
-							isBusinessPlan( planName ) ||
-							isPremiumPlan( planName ) ||
-							isEcommercePlan( planName ),
+						'popular-plan-parent-class': isBusinessPlan( planName ),
 						'plan-is-footer': isFooter,
 					}
 				);
